replaced race-prone event raisings with atomic ones
Previously, the events were being raised by testing them for nullness
and, inside the if block, raising the event. This is a problem for
concurrency because theoretically the nullness could change between
the test and execution.

The ?. operator (null-conditional access) fixes this problem since
it is guaranteed to be an atomic operation.